Bug #640407 fix the ril converity error
[root cause  ] converity resource leak
[changes     ] add the response free,fix the converity error
[side effects] no
[self test   ]<board, Test cases summary>:<9832,boot check the network register>
[reviewers   ] zhiwei.fu
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

+++ sprd-ril/atchannel.c
@@ -916,7 +916,7 @@ long long getATTimeoutMesc(const char *command) {
     for (i = 0; i < s_ATTableSize; i++) {
         if (!strncasecmp(s_ATTimeoutTable[i].cmd, command,
                 s_ATTimeoutTable[i].len)) {
-            timeoutMesc = s_ATTimeoutTable[i].timeout * 1000;
+            timeoutMesc = s_ATTimeoutTable[i].timeout * (long long)1000;
             break;
         }
     }
